Adding intro map
Used in the Intro campaign, adding these files to your addons folder will let you use it as an addon as well. The _blank file has the campaign objective removed, as well as the lift & slopes from the campaign removed. The normal file is exactly how this map appeared in the campaign. Just say you want to "keep playing" when the message pops up on either. There are no objectives.
 master